Remove shared/rational folder, move specs inline
This has become obsolete in 2015 when Rational was imported into core

Reference: commit 3ceb9ae03b6333ab892fec4efabfc5b348f226f0
