+++ core/rational/div_spec.rb
@@ -1,18 +1,54 @@
 require_relative "../../spec_helper"
-require_relative '../../shared/rational/div'
 
 describe "Rational#div" do
-  it_behaves_like :rational_div, :div
+  it "returns an Integer" do
+    Rational(229, 21).div(82).should be_kind_of(Integer)
+  end
+
+  it "raises an ArgumentError if passed more than one argument" do
+    -> { Rational(3, 4).div(2,3) }.should raise_error(ArgumentError)
+  end
+
+  # See http://redmine.ruby-lang.org/issues/show/1648
+  it "raises a TypeError if passed a non-numeric argument" do
+    -> { Rational(3, 4).div([]) }.should raise_error(TypeError)
+  end
 end
 
 describe "Rational#div passed a Rational" do
-  it_behaves_like :rational_div_rat, :div
+  it "performs integer division and returns the result" do
+    Rational(2, 3).div(Rational(2, 3)).should == 1
+    Rational(-2, 9).div(Rational(-9, 2)).should == 0
+  end
+
+  it "raises a ZeroDivisionError when the argument has a numerator of 0" do
+    -> { Rational(3, 4).div(Rational(0, 3)) }.should raise_error(ZeroDivisionError)
+  end
+
+  it "raises a ZeroDivisionError when the argument has a numerator of 0.0" do
+    -> { Rational(3, 4).div(Rational(0.0, 3)) }.should raise_error(ZeroDivisionError)
+  end
 end
 
 describe "Rational#div passed an Integer" do
-  it_behaves_like :rational_div_int, :div
+  it "performs integer division and returns the result" do
+    Rational(2, 1).div(1).should == 2
+    Rational(25, 5).div(-50).should == -1
+  end
+
+  it "raises a ZeroDivisionError when the argument is 0" do
+    -> { Rational(3, 4).div(0) }.should raise_error(ZeroDivisionError)
+  end
 end
 
 describe "Rational#div passed a Float" do
-  it_behaves_like :rational_div_float, :div
+  it "performs integer division and returns the result" do
+    Rational(2, 3).div(30.333).should == 0
+    Rational(2, 9).div(Rational(-8.6)).should == -1
+    Rational(3.12).div(0.5).should == 6
+  end
+
+  it "raises a ZeroDivisionError when the argument is 0.0" do
+    -> { Rational(3, 4).div(0.0) }.should raise_error(ZeroDivisionError)
+  end
 end
