+++ core/rational/fdiv_spec.rb
@@ -1,6 +1,5 @@
 require_relative "../../spec_helper"
-require_relative '../../shared/rational/fdiv'
 
 describe "Rational#fdiv" do
-  it_behaves_like :rational_fdiv, :fdiv
+  it "needs to be reviewed for spec completeness"
 end
