+++ core/rational/fixtures/rational.rb
