+++ core/rational/hash_spec.rb
@@ -1,6 +1,9 @@
 require_relative "../../spec_helper"
-require_relative '../../shared/rational/hash'
 
 describe "Rational#hash" do
-  it_behaves_like :rational_hash, :hash
+  # BUG: Rational(2, 3).hash == Rational(3, 2).hash
+  it "is static" do
+    Rational(2, 3).hash.should == Rational(2, 3).hash
+    Rational(2, 4).hash.should_not == Rational(2, 3).hash
+  end
 end
