+++ core/rational/inspect_spec.rb
@@ -1,6 +1,14 @@
 require_relative "../../spec_helper"
-require_relative '../../shared/rational/inspect'
 
 describe "Rational#inspect" do
-  it_behaves_like :rational_inspect, :inspect
+  it "returns a string representation of self" do
+    Rational(3, 4).inspect.should == "(3/4)"
+    Rational(-5, 8).inspect.should == "(-5/8)"
+    Rational(-1, -2).inspect.should == "(1/2)"
+
+    # Guard against the Mathn library
+    guard -> { !defined?(Math.rsqrt) } do
+      Rational(bignum_value, 1).inspect.should == "(#{bignum_value}/1)"
+    end
+  end
 end
