+++ core/rational/magnitude_spec.rb
@@ -1,5 +1,5 @@
 require_relative "../../spec_helper"
-require_relative '../../shared/rational/abs'
+require_relative 'shared/abs'
 
 describe "Rational#abs" do
   it_behaves_like :rational_abs, :magnitude
