+++ core/rational/quo_spec.rb
@@ -1,6 +1,25 @@
 require_relative "../../spec_helper"
-require_relative '../../shared/rational/divide'
 
 describe "Rational#quo" do
-  it_behaves_like :rational_divide, :quo
+  it "calls #coerce on the passed argument with self" do
+    rational = Rational(3, 4)
+    obj = mock("Object")
+    obj.should_receive(:coerce).with(rational).and_return([1, 2])
+
+    rational.quo(obj)
+  end
+
+  it "calls #/ on the coerced Rational with the coerced Object" do
+    rational = Rational(3, 4)
+
+    coerced_rational = mock("Coerced Rational")
+    coerced_rational.should_receive(:/).and_return(:result)
+
+    coerced_obj = mock("Coerced Object")
+
+    obj = mock("Object")
+    obj.should_receive(:coerce).and_return([coerced_rational, coerced_obj])
+
+    rational.quo(obj).should == :result
+  end
 end
