+++ core/rational/remainder_spec.rb
@@ -1,6 +1,5 @@
 require_relative "../../spec_helper"
-require_relative '../../shared/rational/remainder'
 
 describe "Rational#remainder" do
-  it_behaves_like :rational_remainder, :remainder
+  it "needs to be reviewed for spec completeness"
 end
