+++ core/rational/shared/abs.rb
@@ -1,4 +1,4 @@
-require_relative '../../spec_helper'
+require_relative '../../../spec_helper'
 
 describe :rational_abs, shared: true do
   it "returns self's absolute value" do
