+++ core/rational/to_i_spec.rb
@@ -1,6 +1,12 @@
 require_relative "../../spec_helper"
-require_relative '../../shared/rational/to_i'
 
 describe "Rational#to_i" do
-  it_behaves_like :rational_to_i, :to_i
+  it "converts self to an Integer by truncation" do
+    Rational(7, 4).to_i.should eql(1)
+    Rational(11, 4).to_i.should eql(2)
+  end
+
+  it "converts self to an Integer by truncation" do
+    Rational(-7, 4).to_i.should eql(-1)
+  end
 end
