--- shared/rational/denominator.rb
@@ -1,14 +1,0 @@
-require_relative '../../spec_helper'
-
-describe :rational_denominator, shared: true do
-  it "returns the denominator" do
-    Rational(3, 4).denominator.should equal(4)
-    Rational(3, -4).denominator.should equal(4)
-
-    Rational(1, bignum_value).denominator.should == bignum_value
-  end
-
-  it "returns 1 if no denominator was given" do
-    Rational(80).denominator.should == 1
-  end
-end
