--- shared/rational/fdiv.rb
@@ -1,5 +1,0 @@
-require_relative '../../spec_helper'
-
-describe :rational_fdiv, shared: true do
-  it "needs to be reviewed for spec completeness"
-end
