--- shared/rational/to_r.rb
@@ -1,11 +1,0 @@
-require_relative '../../spec_helper'
-
-describe :rational_to_r, shared: true do
-  it "returns self" do
-    a = Rational(3, 4)
-    a.to_r.should equal(a)
-
-    a = Rational(bignum_value, 4)
-    a.to_r.should equal(a)
-  end
-end
